+++ CHANGELOG.md
@@ -11,6 +11,7 @@ ChangeLog
 - CLEANUP: add license headers
 - CLEANUP: various code cleanups (naming, simplification, resource handling, messages)
 - CLEANUP: various cleanups to the source trait
+- CLEANUP: pass-through specific error on cancelAndUnroll in service trait
 - UPDATE: upgrade NPM dependencies
 
 1.4.17 (2026-04-11)